Add Rubocop for linting
Rubocop provides automation for code style linting.

This enables the default set of rules based on the community maintained
Ruby Style Guide.

These rules aren't set in stone and anyone in the team can suggest
modifications via a PR and get team buy-in for that change.

The goal is to avoid bike-shedding in our PRs that can slow down changes
to the app. If you see a code style in someone's code that you disagree
with, rather than asking them to make the change in their PR, propose a
change to a cop in a separate PR to ensure this gets automated.