Add eslint + prettier
We want to enable linting in our Javascript to avoid bikeshedding
de-railing code changes there.

eslint is a great tool for this. I've opted to use the AirBnB style
guide. They are a team that write a lot of Ruby too and share a similar
approach to writing Ruby + JS apps.

This change also introduces prettier. This is another formatting tool
that helps auto-correct code and has wide support across IDEs and
text editors.

The goal is to reduce the amount of time we spend on code-style and
formatting, both as we write the code and when we review.