Setup system tests
We want a way to be able to run tests in a real browser environment.

[cuprite](https://github.com/rubycdp/cuprite) is a pure Ruby
implementation of CDP (Chrome Developer Protocol). It makes it possible
to run our tests in Chrome without requiring Selenium and the chrome
webdriver.

This leads to an easier setup for developers and also a slightly faster
test suite.