Add tool-versions for asdf
[`asdf`](https://asdf-vm.com/) is a modern runtime version management
tool, which works like `rbenv` or `nvm`. It has a plugin system that
allows extending it to any language.

The plugins are able to parse `.ruby-version` and similar
language-native version definition files, but adding a central
`.tool-versions` file is generally recommended. We should keep these in
sync.

I'll document how to install deps in the README in a follow-up PR.