Add prettier for Ruby formatting
There is a preference within the team to use prettier for formatting
Ruby as well as our JS.

This introduces the plugin on the Ruby side (it is already enabled for
JS) and ensures it doesn't collide with the rules provided by Rubocop.

We get the best of both tools with this setup.

There is a command to make running both tools easier, `bin/lint` runs
prettier on all supported files and then Rubocop with autocorrect
enabled.

Use this if you don't have the tools running in your editor while you
write code.