Setup deployment to PaaS
The PaaS service is the preferred method of deployment for this app.

The latest version of Ruby that is supported on PaaS is 3.0.3, so I had
to modify the app to use that version.

Also, the latest version of Bundler is 2.2.33, so I had to downgrade in
order for the deployment to complete successfully.

We've opted to use GitHub Actions to manage our continuous deployment
process. It's the logical choice given all our CI actions take place
here too.