Add a markdown renderer
We want to add the ability to render markdown templates to make it
easier for people to add and format copy.

There is a DfE-Digital library for a GOVUK styled markdown renderer,
called [GovukMarkdown](https://github.com/DFE-Digital/govuk_markdown).

I opted to register this as a Rails template handler to make it simple
for someone to add a markdown partial and include it in a view template.

The other option I considered was to write a helper method that we could
include in the erb templates that would convert a markdown string into
HTML. Ultimately, this seemed like it would end up with code that was
harder to read.

I think the separation of markdown and HTML makes the most sense from a
developer experience point of view.