Ensure asset directories are present
There's an issue on deployment when precompiling assets.

The folders esbuild expects to be present are hidden by the .gitignore
file.

We did this to prevent locally compiled assets from being checked in to
version control.

What I missed in the initial setup was to allow the presence of those
directories, using a .keep, and only ignoring the contents.