Add a confirmation page
When a person has completed a request for a TRN, we want to display a
confirmation screen that provides details of what to expect next.

The reference design is [here](https://find-a-lost-trn.herokuapp.com/helpdesk-request-submitted).

`TrnRequest` is intended to be a record of the request. Currently, we
have no details to store on it. This can be iterated on as we progress.

For now, we simply use the ID of the record as a reference number for
person who made the request.

I opted to store this ID in the session so that the confirmation screen
has access to it.

This also introduces a way to add a prefix to the page title from a
view.

Simply use the `content_for` method using `:page_title` as the target
and it will get injected into the page title in the <head> of the
document.