Add the Check your answers page
After providing all the details required to find a lost TRN, there is a
screen that displays the answers given and presents an opportunity to
modify them before submitting the request.

The reference design for this screen is [here](https://find-a-lost-trn.herokuapp.com/check-answers) and the
Trello ticket is [here](https://trello.com/c/cVaeihfv).

Assumptions:
* the column types for the fields have been set to what appears to be
  the correct type. This could change as we implement the relevant
  screens.
* the `checked_at` field is a placeholder to represent the submit action
  by the user. This could be replaced by a finite state machine in
  future to better represent the state of the TRN request.
* the back link goes to /start as that also happens to be the previous
  screen. Once we add more of the flow, this will change.
* the TRN request values are hard-coded until we implement the screens
  to capture them.