Add the email page
We want to allow people to enter an email address as part of the TRN
request.

This adds a new page for collecting the email address.

We're able to change the start button on the start page back to a link
as it no longer needs to POST to create the `TrnRequest`.

I'm not sure how long the `TrnRequest` object will continue as we have
talked about using a `Teacher` object for the long term. I opted to
leave the models as they are for now, with the assumption that it is
relatively trivial to change a model name in the future.

The `edit` action for the `EmailController` also takes on the
responsibility of creating the `TrnRequest` object. I expect this to move in
the future to whatever screen precedes it.

`ValidForNotifyValidator` is a direct copy of the validator originally
written in [apply-for-teacher-training](https://github.com/DFE-Digital/apply-for-teacher-training/blob/main/app/validators/valid_for_notify_validator.rb/).

The assumption is that emails will be sent using Notify, so the
validations should line up with the expectations of that system.

I excluded the system specs from the `Rspec/ExampleLength` cop because
the default of 5 lines is restrictive when writing scenarios that have
multiple steps to test the same behaviour across multiple screens.

This seemed better to me than the alternative of writing a separate
scenario for each page, eg. testing the back link.