Add NI number to the request flow
The designs for the TRN request flow have steps to collect an NI number.

This is an optional step and split into 2 screens. The first to
determine if the person knows their NI number and the second to allow
them to enter it.

I tried a different approach to validating the TrnRequest. Rather than
adding a conditional validation to the `TrnRequest` class, I created a
form object, `NiNumber` that gets used to specifically add the
`ni_number` value and makes it easier to conditionally validate that
value.

It means the object representing the record doesn't need to know about
rules related to the flow.

This makes this particular object easier to understand. It comes at the
cost of spreading validation rules across multiple other objects.

At this stage, I think the trade-off is appropriate.

Assumptions:
* we validate the format of the number but have no way of checking if it
  is a real national insurance number.