Add a date of birth question
We want to collect a person's date of birth to help with finding their
lost TRN.

We're following guidance on dates from [this design system
entry](https://design-system.service.gov.uk/patterns/dates/Add a date of
birth question

We want to collect a person's date of birth to help with finding their
lost TRN.

We're following guidance on dates from [this design system entry](https://design-system.service.gov.uk/patterns/dates/.

[Reference design](https://find-a-lost-trn.herokuapp.com/dob) and
[Trello ticket](https://trello.com/c/qTIsoCgM/159-dob-page-date-of-birth-your-date-of-birth).

Validations are handled in a similar fashion to those for the national
insurance number. There is a form object that gets used by the
controller and view. It contains the validation logic and allows the
underlying model, `TrnRequest` to remain free from the complexity of the
order we present the questions in.

Assumptions:
* a valid date of birth should mean the person is at least 16 years old.
* people over 120 years old are unlikely to be using this service.