Ensure the session ID is set correctly
We store the TrnRequest#id in the session to be able to pass state
between screens. The value needs to be set when the first question is
answered.

This change was missed when we added the NI questions and as a result
the TrnRequest doesn't persist correctly between questions.

As we add new questions, this step will continue to migrate to the front
of the flow and we will need to add better tests to ensure we don't miss
this in the future.

For now, I avoided adding extra tests as there are other issues to be
fixed before we can do that.