Validate the NI number correctly
The validations for the NI number fields aren't triggering correctly in
the UI.

The 2 issues are that the error message for the validation is the Rails
default rather than the GOVUK style, which makes identifying the issue
hard. The GOVUK component doesn't render the field name in the errors,
so it relies on the error message including it.

Also, the ITT provider enrolled validation is firing at the wrong time.
This needs to be extracted to its own form object (which will happen in
another PR).