Create new back link helper
This uses a specified URL, or the `:back` Rails action as a fallback
when nothing is specified. When `session[:form_complete]` is set, the
behaviour changes so that the back button takes users back to the form.
We re-set this when users submit the form or go back to the homepage.