Find a TRN using the DQT API
There is an API we can use to find a TRN and confirm the details of a
teacher.

The response we get from this API is an array containing zero or more
matching teacher records.

We don't have a plan to disambiguate multiple matches automatically, so
in those scenarios we will send the request to ZenDesk for human
intervention.

If there is a single match, then we will treat this as if it is an exact
match and display the returned TRN.

This assumption (a single result in the API response is an exact match)
might not be true. If this is the case, then we may want to revisit this
in the future.

I introduced VCR in this change in order to make our tests
deterministic. The pre-prod environment that we hit in our tests uses an
anonymised copy of the production DB.

It can also be reset at arbitrary intervals.

As a result, we can't guarantee that the data would be the same across
test runs.

VCR allows us to cache a response and re-run the tests against that
response until we decide to update the cache.

An assumption here is that we will only update this cache if we get
notified that the API has changed.