Enable logging on calls to external APIs
We don't currently log the calls made to external APIs, (only DQT at the
moment). This makes it harder to debug when things don't work as
expected.

I enabled the logger middleware in Faraday and opted to include the
response headers and body for full insight into the response.

There is little value in knowing the specific values of the params
passed to the API call, especially when compared to the risk of
accidentally leaking PII. For this reason, we are filtering all PII from
these logs. This also ensures compliance with GDPR.