Fail loudly when Zendesk integration is off
In https://github.com/DFE-Digital/find-a-lost-trn/pull/137 we introduced
a dependency on displaying `.zendesk_ticket_id` in the success page
after a user submits a TRN request. There's no way to obtain this ID
without either connecting to the dummy Zendesk client (in development)
or the production one. So we should now fail loudly if we attempt to
submit a TRN request without the feature flag being on.