Fix the back link for the itt-provider screen
When the 'Do you have a National Insurance number?' question has been
answered in the affirmative, the back link for the ITT provider screen
should return to the form to enter a national insurance number.
Currently, it goes to the screen previous to that.

I opted to check the answer in the itt provider view to determine the
correct URL to go back to. That seemed to be the simplest and easiest to
understand solution here.