Fix back link for the check answers screen
When viewing the check answers path, the back link
always points to itself.

There's a bug in the helper method that builds this link. When the form
is marked as complete, it always goes to the check answers screen.

I've updated it to ignore this if the link is generated on the check
answers screen.