Add a presence validation for email
Currently, we use the notify format validations to also validate the
presence of an email address. This gives a slightly strange error
message for the nil case.

Our preference is to make that nil error simpler and notify the user
that the email address is missing.

I opted to conditionally apply the notify validations to avoid
displaying multiple error messages when the email field is blank.