Update the request submitted screen
A couple of small copy changes here.

* We want to ensure that the title keeps 'request received' on the same line.
* There's a copy change relating to instructions about urgent requests.

In order to get the title displaying correctly, I needed to insert a
non-breaking space character between the words we want to keep together.

The view component we use for the panel doesn't have an attribute to
allow inserting anything but plain text. As a result, I needed to
fallback to writing the HTML for the header.