Fix NI edge case when displaying anonymised data
We had this bug happen in Sentry: https://sentry.io/organizations/dfe-teacher-services/issues/3207003600/?project=6275068&referrer=slack

It's because we attempted to call `.first` on a nil NI number on a
TRN request that had answered yes to "Do you have a NI number?" but then
didn't actually fill the number in.

To work around this, we can add a couple of guard clauses.