Add a question about being awarded a QTS
As part of the lessons learned from user feedback, we want to separate
out the question of whether someone has been awarded QTS from how they
were awarded it.

This allows us to be more inclusive of people who were awarded a QTS outside a
school or university.

I opted to introduce a new attribute on `TrnRequest` to store the answer
to the question of if someone has a QTS.

This allows us to use the `itt_provider_enrolled` attribute to now mean
whether someone was able to supply the name of the training provider or
not.