Add logging for validation errors
We want visibility into how people are using the service and in
particular, which error messages they see.

I took inspiration from [Apply](https://github.com/DFE-Digital/apply-for-teacher-training/pull/1922) when designing this.

Storing validation errors as a record in the database makes it easy for
us to display in the support interface. The UI for this will come in a
future change.

The scope of where a validation error can occur is more limited than the
Apply version, so I was able to keep it simpler.

An error will always be associated with a `TrnRequest` and the form
object that was validated has a one-to-one mapping with a particular
path, so we know where the error was seen.

I considered using a controller concern but instead went for including
it at the form object level. The idea being that logging the errors is
the responsibility of the form object that triggered the errors rather
than having another object have knowledge of both the form object and
how to record the errors.

Apply made the choice to log in the controller to avoid logging
validation errors more than necessary, ie. they assumed that form
objects would be used in more places than the controller.

In our case, we have an assumption that a form object is only used
by the controllers. We have no other way to use them. The pattern we
have followed is to use the form objects to encapsulate all side-effects
related to creating/mutating objects in the system.

This means that controller actions are responsible for co-ordinating
this process but not the implementation of it.