Schedule performance metrics to post to Slack
We want to provide an easy way to track the performance metrics from
Find a Lost TRN.

It was decided that a message in Slack was the most appropriate way of
doing this.

We followed the lead of the Apply project to implement the scheduling
piece of this work with [Clockwork](https://github.com/Rykian/clockwork).

Right now, the message we post is very simple and only contains the
total number of TrnRequests created in the last week.

To complete setup of this, a Slack bot authorisation token is required,
ie. SLACK_TOKEN, and also the channel it posts to is customisable,
SLACK_CHANNEL.

These configuration variables allow us to run this in staging and verify
it works.