Allow people to continue without a national insurance number
When someone indicates they have a national insurance number, we present
them an option to skip entry if they don't know it.

The way we enforce the question order means that no-one is able to skip
this step and is always redirected back to the national insurance form.

We are relaxing constraints that prevent people from skipping the
national insurance question by typing in the ITT provider URL.

This allows us to enable skipping via the 'Continue without it' link.