Use a Slack webhook
After feedback on getting access to a new Slack user to enable API
access, we have decided to switch to a simpler method of posting a
message.

Using a Slack webhook simplifies things. We create a webhook URL that is
specific to a channel in Slack. Sending a POST request to that URL
results in a message being added to the channel.

No further authentication is required.