Enable retrieving data from Zendesk
When a TRN Request cannot be found via the API, a Zendesk ticket is
created. Often, this manual intervention results in a matching TRN being
found.

We want to be able to retrieve this TRN from the Zendesk ticket and then
use it to lookup the details of the teacher via the API and then create
a diff of the details provided with the details returned by the API.

The aim is for us to be able to learn how we can adapt the service to
resolve these scenarios without a manual intervention.

The first pass at this implements a mechanism to retrieve the TRN from
Zendesk and also retrieve the response from the DQT API. However, it is
required to be triggered manually by pressing a button in the support
interface.

The automation of this process can be implemented in a future change.