Display the DQT data consistently with the request data
Displaying the diff differently to the request data makes it harder to
recognise that this is the same kind of data.

As a solution, we can render the DQT data as rows in the existing list
of request data, differentiated with a prefix, DQT.

Using it here also means we should follow the convention of anonymising
data in production. The reasoning here is that in production, no-one
needs to see the exact values that are different. The knowledge that the
fields are different is enough for our purposes.

This new view component for the TrnResponse looks quite similar to the
existing one for TrnDetails. I chose not to de-duplicate and make the
TrnDetails component more generic to handle both cases. This feels like
it can handled by a follow-up change.