Enable iterative processing of Zendesk comments
When searching for a TRN in the comments of a Zendesk ticket there is an
assumption that the last comment contains the matching text.

In practice, this isn't always true. A ticket can have multiple comments
and the last one isn't guaranteed to contain the TRN.

This change will iteratively process the comments, starting from the
second one (the first is always a copy of the original request). The
search will stop as soon as the regex matches.

I opted for this as the simplest improvement we could make to the
current process. It still makes an assumption that the regex will
eventually match. This assumption may not always be true in practice.

We are looking into ways of making this more reliable in the future.