Fix error with multiple comments on a ticket
When a Zendesk ticket has multiple comments and none of them contain a
match for a TRN, the service throws an `NoMethodError`.

It expects there to always be a match when there are more than one
comment on a ticket. This is a faulty assumption.

Instead, we will only attempt to retrieve a TRN from a comment if there
has been a match first.