Simplify no TRN content
We are removing instructions on how to get a TRN for NPQ users – they will have those instructions in the NPQ registration journey.

Instead we are matching the guidance we show on GOV.UK – users who are eligible for a TRN get given one, they should not need to do anything.