Customise the performance alert URL per host
The URL that is included in the performance alert is hard-coded to
production. This makes it hard to distinguish alerts from other
environments.

We occasionally need to test this alert in other environments, so it is
beneficial to allow us to customise the URL.

I opted to use an environment variable to store the current hosting
domain, that can then be passed to the Rails URL helper to generate the
appropriate value.

This seems simpler and easier to extend across environments than the
alternative of keeping a list of all expected domains in code somewhere.