Queue a job to periodically check Zendesk for a TRN
When a Zendesk ticket gets created, we have a manual trigger to check
if a TRN has been found and sent to the requesting person.

We want to automate this process.

I investigated using a Zendesk webhook but struggled to be able to get
them set up correctly. I believe this will be a better way but in the
short term, polling the ticket will be an improvement on the manual
process.