Handle a Zendesk API failure
We had an incident where the connection to the Zendesk API failed and
we were unable to create tickets for requests that had no match.

This resulted in failed requests and manual work to resolve the issue.

We discussed solutions for this and decided that using our job queue
would allow us to gracefully handle this and retry the requests to
Zendesk at a later time to ensure they still get created.

If the API fails, we inform the person requesting their TRN that there
will be a delay while we address the issue. Then we queue a job to keep
trying to create the ticket.

The default Sidekiq retry schedule ensures that the job will be retried
at exponentially increasing intervals. This should give the API time to
recover or, in the case where manual intervention is required, we have
a mechanism to retry the job manually.