Add a daily check for the Zendesk connection
A recent incident highlighted that we don't have a good alerting process
around the connection to the Zendesk API.

We decided to add a daily check to make sure the connection is working.

This solution isn't a definitive solution, but it is a start. We are
relying on the presence of a TRN request with a Zendesk ticket ID as a
proxy for the health of the connection.

The assumption is that if we have been able to create at least one
one ticket in the past 24 hours, then the connection is working.