Improve the health check endpoint
We want to add more checks to the /health endpoint to ensure that it
represents the true health of the system.

The 'ok_computer' gem provides a simple set of checks and a middleware
that we can mount at `/health`.

This change uses the checks that come with the gem. One for each
datastore we are using plus a Sidekiq latency check.

I left the settings as the defaults but we can tweak if we find they are
not suitable for our system.

We can use this check framework to write custom checks for Zendesk and
the DQT API. I propose adding these extra checks in subsequent changes.