Use form submit for Continue without NI
We previously used a link to handle the case where users don't know
their NI number. The downside to this is that a link performs a GET
request, and we shouldn't mutate any server-side state via GETs.

The current approach for handling the "Continue without it" works around
this limitation in the `EnforceQuestionOrder` concern, but we're
changing the way that works to make it easier to reorder steps.

This commit switches the "Continue without it" button to instead use a
`govuk_submit` with a specific `name` and `value`. When the value is
`ni_number_not_known`, we can fork in the controller and redirect the
user to the appropriate path.

Because submitting a form with the keyboard (Alt+Enter) will default to
submitting it via the first Submit button encountered, I had to add
another hidden submit button earlier in the element order. I labelled it
"Submit" instead of "Continue" to make it distinct for Capybara. It's
visually hidden by default, and it's also not a part of the tab order.

I added `name/value`s to all 3 of the buttons as such:

- "Submit": `submit => hidden_submit`
- "Continue without it": `submit => ni_number_not_known`
- "Continue": `submit => submit`