Capture Zendesk error details
When we receive an error message from Zendesk, we don't pass it on to
Sentry, which makes debugging harder.

The exception that gets raised has an `errors` method that contains the
details of the error that has just occurred. I opted to pass this on to
Sentry using the context feature. This will display the details of the
error in the Sentry UI.