Performance page: fix how abandonments and no-matches are counted
After this change, an abandonment is counted whenever the TRN hasn't been found
and there's no Zendesk ticket.

This commit also updates the spec to use traits so the test data
is a bit clearer. The test case itself is updated to count
no-matches instead of abandonments, as that's going to be more
common.