Display a link to Zendesk when present
If a TRN has been found for a request, we don't show a Zendesk ticket
link. This can lead to an incorrect assumption that a ticket wasn't
created.

It is useful to know about a Zendesk ticket even when the TRN has been
found, so we're opting to display the link in the support interface
whenever `zendesk_ticket_id` is set.