Avoid loading Rails for the smoke test
There's an issue when the smoke test is called in a hosted environment,
it tries to load Rails but fails.

We don't want to load the application in the smoke test as we are
trying to test the hosted version.

This change switches the smoke test to not load the application.