Performance page: fix double counting in TRN request classification
This commit fixes a bug in the "TRN requests by day" and the
"TRN requests by month" sections.

For a TRN request that was initially unsuccessful, a Zendesk ticket ID
will be populated. Later, when the helpdesk finds the TRN for this user,
the TRN will also be populated on the record via the background Zendesk
scanning job.

These requests were being double-counted in both the "TRN found" and
"No match" categories.