Allow users to not specify a previous name
Adds an extra option to the name form to allow users that have had a previous name to specify that they prefer to not add a previous name.

Includes:

Installation of the data-migrate gem for data migrations.
A data migration to update the status of a trn requests name change state