Use Sidekiq to deliver emails
We want to send emails using a background job to make them retryable in
the case of failure.

I considered writing a custom job to handle this but opted instead to
keep things simple and use the default setup.

Calling `deliver_later` queues a `ActionMailer::MailDeliveryJob` to
perform sending the email.

I opted to keep using the `default` queue for these jobs as we don't
have a large volume of jobs where separate queues would help with
prioritisation.