Cache ITT provider results for 1 hour
The DQT API returns results fairly quickly; a few hundred milliseconds.
However, since users are likely to hit this endpoint frequently, this
could add quite a bit of strain to the API.

The results are fairly static, so caching them for an hour pre-emptively
seems sensible and not a premature optimisation.