Add an OKComputer check for Zendesk
We want to include the Zendesk integration in the system health checks.

The system is healthy if we can successfully connect to the Zendesk API.
This covers scenarios where we have inadvertently mis-configured Zendesk
and prevents these kind of errors going unnoticed until we get an error
flagged in Sentry.

I opted to fetch the latest count of tickets from the Zendesk API, with
the rationale that it doesn't require needing to know a specific ID of a
ticket that may or may not exist.

Also, it is a value that appears to be cached by Zendesk, so is
potentially the fastest endpoint we can call.