Fix unlock teacher method
Our test stubs were using a hash that contains symbols as keys instead
of strings. We were also swallowing the KeyError on the assumption that
the API wouldn't always return `uid`s, but that assumption isn't
correct.

This updates the test stubs and removes the old test.

Paired with @ransom4real.