Add encryption to PII fields on a TRN request
Rails 7 has a feature that can encrypt DB fields.

We want to use this to protect PII stored on a `TrnRequest`.

Doing this incrementally will ensure we can handle issues that arise
from encrypting the fields in manageable way.

This change handles encrypting names. Decryption is handled
automatically by Rails, so no change required there.

Of note, strings are converted to base64 and require more space in the
DB. The recommendation for the default string type is 510 bytes. This
should be enough for strings that are 255 characters in plaintext.