Do not run docker job for production deploys using a non-main branch
The GitHub production deployment jobs need the docker job to be completed prior to starting the deployment. The ‘needs’ option does not check if job conditionals have been met before running it jobs. This commit adds a new docker job that can only be run if using the main branch.