Filter ENV variables scrubbed
Ensure ENV variables scrubbed are in the list of Rails parameter filter as different development environments will have ENV vars we do not want to mess with and it makes for a cleaner scrub. Also it adds a level of standardizing how we name sensitive ENV vars so they can be scrubbed in the future.