Remove the feature flag use_dqt_api
We added a feature flag to control using the DQT API to retrieve a TRN.

This was introduced at a time when we were building out the
functionality.

The flag has been enabled for enough time to give us confidence that we
wouldn't want to turn this feature off in the future.