Add unique encryption keys for production
We want to limit the impact of leaking our encryption keys by having a
separate key for the production environment.

Rails has a credentials file that is encrypted using a master key.
Currently, we share this master key across all our environments, eg.
development, test, production.

It is stored in multiple places, eg. local machines, GitHub Actions and
Dependabot. This increases the attack surface and makes the production
environment more vulnerable.

To reduce the impact of the master key getting leaked, we want to have a
separate credentials file and master key just for the production
environment.

We will store the master key in the keyvault and it will never need to
be used in a local development environment or any third party, such as
GitHub.

Rails provides support for this by passing an `--environment production`
flag to the `bin/rails credentials:edit` command.

For a smooth deployment of this change, we need to change the value of
`RAILS_MASTER_KEY` in the production keyvault prior to shipping this.