Record exceptions in Sentry
When calling the DQT API from both the date of birth and NINO steps,
there is a rescue statement that silently swallows a mix of error and
info exceptions.

I think it would be useful to surface the error exceptions in Sentry so
that we have visibility into errors with the 3rd parties we rely on.